requirements.txt
Flask
mysql-connector-python
Werkzeug
gspread
google-auth
gunicorn
